Add HA response envelope, standby 409 gating, --group flag, lease reset, and deterministic startup role
