+++ ha/membership.go
@@ -117,6 +117,15 @@ func JoinMembership(ctx context.Context, target *mongo.Client, opts MembershipOp
 // InstanceID returns this instance's identifier.
 func (m *Membership) InstanceID() string { return m.instanceID }
 
+// Host returns the advertised host for this instance.
+func (m *Membership) Host() string { return m.host }
+
+// Port returns the advertised HTTP port for this instance.
+func (m *Membership) Port() int { return m.port }
+
+// Group returns the HA group name this instance joined.
+func (m *Membership) Group() string { return m.group }
+
 // SetRole updates the role and term this instance advertises in its member
 // document. It is the one-way feed from the election layer (the lease document
 // is authoritative): the elected role is pushed here, and the heartbeat loop
@@ -241,6 +250,7 @@ func (m *Membership) beat(ctx context.Context) error {
 	// Aggregation-pipeline update so $$NOW resolves on the server.
 	update := mongo.Pipeline{
 		{{"$set", bson.D{
+			{fieldGroup, m.group},
 			{"host", m.host},
 			{"port", m.port},
 			{"role", role},
